+++ weblate/formats/ttkit.py
@@ -56,6 +56,7 @@ from weblate.formats.base import (
 )
 from weblate.lang.data import FORMULA_WITH_ZERO, ZERO_PLURAL_TYPES
 from weblate.lang.models import Plural
+from weblate.trans.file_format_params import get_encoding_param
 from weblate.trans.util import (
     get_clean_env,
     get_string,
@@ -259,7 +260,6 @@ class TTKitFormat(TranslationFormat):
     set_context_bilingual = True
     # Use settarget/setsource to set language as well
     use_settarget = False
-    force_encoding: str | None = None
     plural_preference: tuple[int, ...] | None = (
         Plural.SOURCE_CLDR,
         Plural.SOURCE_DEFAULT,
@@ -274,13 +274,16 @@ class TTKitFormat(TranslationFormat):
 
     def fixup(self, store) -> None:
         """Perform optional fixups on store."""
-        if self.force_encoding is not None:
-            store.encoding = self.force_encoding
+        if (encoding := self.get_encoding()) is not None and encoding != "auto":
+            store.encoding = encoding
         # This gets already native language code, so no conversion is needed
         if self.language_code is not None:
             store.settargetlanguage(self.language_code)
         if self.source_language is not None:
             store.setsourcelanguage(self.source_language)
+
+    def get_encoding(self) -> str | None:
+        return get_encoding_param(self.file_format_params)
 
     def load(
         self,
@@ -301,17 +304,25 @@ class TTKitFormat(TranslationFormat):
         return store
 
     @classmethod
-    def get_class(cls) -> TranslationStore:
+    def get_class(cls, encoding: str | None = None) -> TranslationStore:
         """Return class for handling this module."""
         # Direct class
         if inspect.isclass(cls.loader):
             return cls.loader
-        # Tuple style loader, import from translate toolkit
-        module_name, class_name = cls.loader
+
+        if isinstance(cls.loader, dict):
+            # With encoding variants
+            if encoding in cls.loader:
+                module_name, class_name = cls.loader[encoding]
+            else:
+                module_name, class_name = next(iter(cls.loader.values()))
+        else:
+            # Tuple style loader, import from translate toolkit
+            module_name, class_name = cls.loader
+
         if "." not in module_name:
             module_name = f"translate.storage.{module_name}"
         module = importlib.import_module(module_name)
-
         # Get the class
         return getattr(module, class_name)
 
@@ -323,7 +334,7 @@ class TTKitFormat(TranslationFormat):
 
     def get_store_instance(self, **kwargs):
         kwargs.update(self.get_format_class_kwargs())
-        store = self.get_class()(**kwargs)
+        store = self.get_class(get_encoding_param(self.file_format_params))(**kwargs)
 
         # Apply possible fixups
         self.fixup(store)
@@ -469,8 +480,8 @@ class TTKitFormat(TranslationFormat):
                 unit.untranslate(language)
 
     @classmethod
-    def get_new_file_content(cls):
-        result = cls.new_translation
+    def get_new_file_content(cls, encoding: str | None = None):
+        result = cls.get_new_translation(encoding)
         if isinstance(result, str):
             result = result.encode()
         return result
@@ -492,8 +503,10 @@ class TTKitFormat(TranslationFormat):
                 callback(store)
             store.untranslate_store(language)
             store.store.savefile(filename)
-        elif cls.new_translation is not None:
-            Path(filename).write_bytes(cls.get_new_file_content())
+        elif cls.empty_file_template is not None:
+            Path(filename).write_bytes(
+                cls.get_new_file_content(get_encoding_param(file_format_params))
+            )
         else:
             msg = "Not supported"
             raise ValueError(msg)
@@ -511,7 +524,7 @@ class TTKitFormat(TranslationFormat):
         if not base:
             if cls.create_empty_bilingual:
                 return True
-            return monolingual and cls.new_translation is not None
+            return monolingual and cls.empty_file_template is not None
         try:
             if not fast:
                 cls(base, file_format_params=file_format_params)
@@ -1237,7 +1250,7 @@ class PoFormat(BasePoFormat, BilingualUpdateMixin):
     unit_class = PoUnit
 
     @classmethod
-    def get_new_file_content(cls) -> bytes:
+    def get_new_file_content(cls, encoding: str | None = None) -> bytes:  # noqa: ARG003
         """Empty PO file content."""
         return b""
 
@@ -1304,7 +1317,7 @@ class PoMonoFormat(BasePoFormat):
     format_id = "po-mono"
     monolingual = True
     autoload: tuple[str, ...] = ()
-    new_translation = (
+    empty_file_template = (
         'msgid ""\n'
         'msgstr "X-Generator: Weblate\\n'
         "MIME-Version: 1.0\\n"
@@ -1347,7 +1360,7 @@ class XliffFormat(TTKitFormat):
     unit_class = XliffUnit
     language_format = "bcp"
     use_settarget = True
-    new_translation = """<?xml version="1.0" encoding="UTF-8"?>
+    empty_file_template = """<?xml version="1.0" encoding="UTF-8"?>
 <xliff xmlns="urn:oasis:names:tc:xliff:document:1.2" version="1.2">
   <file original="Weblate" source-language="en" datatype="plaintext">
     <body>
@@ -1406,7 +1419,7 @@ class Xliff2Format(XliffFormat):
     format_id = "xliff2"
     loader = Xliff2File
     autoload: tuple[str, ...] = ()
-    new_translation = None
+    empty_file_template = None
     monolingual = False
 
     @staticmethod
@@ -1447,54 +1460,36 @@ class PropertiesBaseFormat(TTKitFormat):
 
 class StringsFormat(PropertiesBaseFormat):
     # Translators: File format name
-    name = gettext_lazy("iOS strings (UTF-16)")
+    name = gettext_lazy("iOS strings")
     format_id = "strings"
-    loader = ("properties", "stringsfile")
-    new_translation: str | bytes | None = "\n".encode("utf-16")
+    empty_file_template: str | bytes | None = "\n"
     autoload: tuple[str, ...] = ("*.strings",)
     language_format = "bcp"
+    loader: ClassVar[dict[str, tuple[str, str]]] = {
+        "utf-8": ("properties", "stringsutf8file"),
+        "utf-16": ("properties", "stringsfile"),
+    }
 
-
-class StringsUtf8Format(StringsFormat):
-    # Translators: File format name
-    name = gettext_lazy("iOS strings (UTF-8)")
-    format_id = "strings-utf8"
-    loader = ("properties", "stringsutf8file")
-    new_translation = "\n"
-
-
-class PropertiesUtf8Format(PropertiesBaseFormat):
-    # Translators: File format name
-    name = gettext_lazy("Java Properties (UTF-8)")
-    format_id = "properties-utf8"
-    loader = ("properties", "javautf8file")
-    new_translation = "\n"
-    language_format = "linux"
-    check_flags = ("auto-java-messageformat",)
-
-
-class PropertiesUtf16Format(PropertiesBaseFormat):
-    # Translators: File format name
-    name = gettext_lazy("Java Properties (UTF-16)")
-    format_id = "properties-utf16"
-    loader = ("properties", "javafile")
-    language_format = "linux"
-    new_translation = "\n"
-    # Translate Toolkit autodetection might fail in some cases.
-    force_encoding = "utf-16"
+    @classmethod
+    def get_new_translation(cls, encoding: str | None = None):
+        if encoding == "utf-16":
+            return "\n".encode("utf-16")
+        return cls.empty_file_template
 
 
 class PropertiesFormat(PropertiesBaseFormat):
     # Translators: File format name
-    name = gettext_lazy("Java Properties (ISO 8859-1)")
+    name = gettext_lazy("Java Properties")
     format_id = "properties"
-    loader = ("properties", "javafile")
+    loader: ClassVar[dict[str, tuple[str, str]]] = {
+        "utf-16": ("properties", "javafile"),
+        "iso-8859-1": ("properties", "javafile"),
+        "utf-8": ("properties", "javautf8file"),
+    }
     language_format = "linux"
-    new_translation = "\n"
+    empty_file_template = "\n"
     autoload: tuple[str, ...] = ("*.properties",)
-    # Java properties need to be ISO 8859-1, but Translate Toolkit converts
-    # them to UTF-8.
-    force_encoding = "iso-8859-1"
+    check_flags = ("auto-java-messageformat",)
 
 
 class JoomlaFormat(PropertiesBaseFormat):
@@ -1503,26 +1498,24 @@ class JoomlaFormat(PropertiesBaseFormat):
     format_id = "joomla"
     loader = ("properties", "joomlafile")
     monolingual = True
-    new_translation = "\n"
+    empty_file_template = "\n"
     autoload: tuple[str, ...] = ("*.ini",)
 
 
-class GWTFormat(StringsFormat):
+class GWTFormat(PropertiesBaseFormat):
     # Translators: File format name
-    name = gettext_lazy("GWT properties (UTF-8)")
+    name = gettext_lazy("GWT properties")
     format_id = "gwt"
-    loader = ("properties", "gwtfile")
-    new_translation = "\n"
+    loader: ClassVar[dict[str, tuple[str, str]]] = {
+        "utf-16": ("properties", "gwtfile"),
+        "iso-8859-1": ("properties", "gwtfile"),
+        "utf-8": ("properties", "gwtutf8file"),
+    }
+    empty_file_template = "\n"
+    autoload: tuple[str, ...] = ()
     check_flags = ("auto-java-messageformat",)
     language_format = "linux"
     supports_plural: bool = True
-
-
-class GWTISOFormat(GWTFormat):
-    # Translators: File format name
-    name = gettext_lazy("GWT properties (ISO-8859-1)")
-    format_id = "gwt-iso"
-    force_encoding = "iso-8859-1"
 
 
 class PhpFormat(TTKitFormat):
@@ -1530,7 +1523,7 @@ class PhpFormat(TTKitFormat):
     name = gettext_lazy("PHP strings")
     format_id = "php"
     loader = ("php", "phpfile")
-    new_translation = "<?php\n"
+    empty_file_template = "<?php\n"
     autoload: tuple[str, ...] = ("*.php",)
     unit_class = PHPUnit
 
@@ -1561,7 +1554,7 @@ class RESXFormat(TTKitFormat):
     loader = RESXFile
     monolingual = True
     unit_class = RESXUnit
-    new_translation = RESXFile.XMLskeleton
+    empty_file_template = RESXFile.XMLskeleton
     autoload: tuple[str, ...] = ("*.resx",)
     language_format = "bcp"
     supports_plural: bool = True
@@ -1575,7 +1568,7 @@ class AndroidFormat(TTKitFormat):
     loader = ("aresource", "AndroidResourceFile")
     monolingual = True
     unit_class = AndroidUnit
-    new_translation = (
+    empty_file_template = (
         '<?xml version="1.0" encoding="utf-8"?>\n<resources>\n</resources>'
     )
     autoload: tuple[str, ...] = ("strings*.xml", "values*.xml")
@@ -1625,7 +1618,7 @@ class JSONFormat(DictStoreFormat):
     loader = JsonFile
     unit_class = JSONUnit
     autoload: tuple[str, ...] = ("*.json",)
-    new_translation = "{}\n"
+    empty_file_template = "{}\n"
     set_context_bilingual = False
 
     @staticmethod
@@ -1682,7 +1675,7 @@ class GoI18JSONFormat(JSONFormat):
     format_id = "go-i18n-json"
     loader = ("jsonl10n", "GoI18NJsonFile")
     autoload: tuple[str, ...] = ()
-    new_translation = "[]\n"
+    empty_file_template = "[]\n"
     supports_plural: bool = True
 
 
@@ -1752,7 +1745,7 @@ class TOMLFormat(TTKitFormat):
     loader = ("toml", "TOMLFile")
     autoload: tuple[str, ...] = ("*.toml",)
     monolingual = True
-    new_translation = "\n"
+    empty_file_template = "\n"
 
     @staticmethod
     def mimetype() -> str:
@@ -1782,7 +1775,6 @@ class CSVFormat(TTKitFormat):
     loader = ("csvl10n", "csvfile")
     unit_class = CSVUnit
     autoload: tuple[str, ...] = ("*.csv",)
-    force_encoding = "auto"
 
     def __init__(
         self,
@@ -1880,20 +1872,11 @@ class CSVFormat(TTKitFormat):
         return result
 
 
-class CSVUtf8Format(CSVFormat):
-    # Translators: File format name
-    name = gettext_lazy("CSV file (UTF-8)")
-    format_id = "csv-utf-8"
-    autoload: tuple[str, ...] = ()
-    force_encoding = "utf-8"
-
-
 class CSVSimpleFormat(CSVFormat):
     # Translators: File format name
     name = gettext_lazy("Simple CSV file")
     format_id = "csv-simple"
     autoload: tuple[str, ...] = ("*.txt",)
-    force_encoding = "auto"
 
     @staticmethod
     def extension() -> str:
@@ -1908,22 +1891,6 @@ class CSVSimpleFormat(CSVFormat):
         return self.parse_simple_csv(content, filename)
 
 
-class CSVSimpleFormatISO(CSVSimpleFormat):
-    # Translators: File format name
-    name = gettext_lazy("Simple CSV file (ISO-8859-1)")
-    format_id = "csv-simple-iso"
-    force_encoding = "iso-8859-1"
-    autoload: tuple[str, ...] = ()
-
-
-class CSVUtf8SimpleFormat(CSVSimpleFormat):
-    # Translators: File format name
-    name = gettext_lazy("Simple CSV file (UTF-8)")
-    format_id = "csv-simple-utf-8"
-    force_encoding = "utf-8"
-    autoload: tuple[str, ...] = ()
-
-
 class YAMLFormat(DictStoreFormat):
     # Translators: File format name
     name = gettext_lazy("YAML file")
@@ -1931,7 +1898,7 @@ class YAMLFormat(DictStoreFormat):
     loader = ("yaml", "YAMLFile")
     unit_class = MonolingualSimpleUnit
     autoload: tuple[str, ...] = ("*.pyml",)
-    new_translation = "{}\n"
+    empty_file_template = "{}\n"
 
     @staticmethod
     def mimetype() -> str:
@@ -1960,7 +1927,7 @@ class DTDFormat(TTKitFormat):
     loader = ("dtd", "dtdfile")
     autoload: tuple[str, ...] = ("*.dtd",)
     unit_class = MonolingualSimpleUnit
-    new_translation = "\n"
+    empty_file_template = "\n"
     can_add_unit: bool = False
 
     @staticmethod
@@ -2045,7 +2012,7 @@ class FlatXMLFormat(TTKitFormat):
     loader = ("flatxml", "FlatXMLFile")
     monolingual = True
     unit_class = FlatXMLUnit
-    new_translation = '<?xml version="1.0" encoding="utf-8"?>\n<root></root>'
+    empty_file_template = '<?xml version="1.0" encoding="utf-8"?>\n<root></root>'
 
     def get_format_class_kwargs(self):
         return {
@@ -2068,7 +2035,7 @@ class ResourceDictionaryFormat(FlatXMLFormat):
     loader = ("resourcedictionary", "ResourceDictionaryFile")
     check_flags = ("c-sharp-format",)
     language_format = "bcp_legacy"
-    new_translation = """<ResourceDictionary
+    empty_file_template = """<ResourceDictionary
   xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
   xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
   xmlns:system="clr-namespace:System;assembly=mscorlib">
@@ -2082,7 +2049,7 @@ class INIFormat(TTKitFormat):
     loader = ("ini", "inifile")
     monolingual = True
     unit_class = INIUnit
-    new_translation = "\n"
+    empty_file_template = "\n"
 
     @staticmethod
     def mimetype() -> str:
@@ -2176,7 +2143,7 @@ class XWikiPropertiesFormat(PropertiesBaseFormat):
     loader = ("properties", "xwikifile")
     language_format = "linux"
     autoload: tuple[str, ...] = ("*.properties",)
-    new_translation = "\n"
+    empty_file_template = "\n"
     can_add_unit: bool = False
     can_delete_unit: bool = False
     set_context_bilingual: bool = True
@@ -2230,7 +2197,6 @@ class XWikiPagePropertiesFormat(XWikiPropertiesFormat):
     name = "XWiki Page Properties"
     format_id = "xwiki-page-properties"
     loader = ("properties", "XWikiPageProperties")
-    force_encoding = "utf-8"
 
     def save_content(self, handle) -> None:
         if self.store.root is None:
@@ -2333,7 +2299,7 @@ class TBXFormat(TTKitFormat):
     format_id = "tbx"
     loader = tbxfile
     autoload: tuple[str, ...] = ("*.tbx",)
-    new_translation = """<?xml version="1.0"?>
+    empty_file_template = """<?xml version="1.0"?>
 <!DOCTYPE martif PUBLIC "ISO 12200:1999A//DTD MARTIF core (DXFcdV04)//EN" "TBXcdv04.dtd">
 <martif type="TBX">
     <martifHeader>
@@ -2378,11 +2344,12 @@ class TBXFormat(TTKitFormat):
         self.store.addheader()
 
 
-class PropertiesMi18nFormat(PropertiesUtf8Format):
+class PropertiesMi18nFormat(PropertiesBaseFormat):
     # Translators: File format name
     name = gettext_lazy("@draggable/i18n lang file")
     format_id = "mi18n-lang"
-    new_translation = "\n"
+    empty_file_template = "\n"
+    loader = ("properties", "javautf8file")
     language_format = "bcp_legacy"
     check_flags = ("es-format",)
     monolingual = True
@@ -2415,7 +2382,7 @@ class StringsdictFormat(DictStoreFormat):
     loader = ("stringsdict", "StringsDictFile")
     unit_class = MonolingualSimpleUnit
     autoload: tuple[str, ...] = ("*.stringsdict",)
-    new_translation = """<?xml version="1.0" encoding="UTF-8"?>
+    empty_file_template = """<?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
 <plist version="1.0">
     <dict>
@@ -2485,7 +2452,7 @@ class FluentFormat(TTKitFormat):
     loader = ("fluent", "FluentFile")
     unit_class = FluentUnit
     autoload: tuple[str, ...] = ("*.ftl",)
-    new_translation = ""
+    empty_file_template = ""
     language_format: str = "bcp"
     check_flags = (
         "fluent-source-syntax",