+++ weblate/trans/tests/test_file_format_params.py
@@ -48,9 +48,7 @@ class BaseFileFormatsTest(ViewTestCase):
         self.assertContains(response, "Settings saved")
         self.component.refresh_from_db()
 
-
-class ComponentFileFormatsParamsTest(BaseFileFormatsTest):
-    def client_create_component(self, result, **kwargs):
+    def client_create_component(self, result: bool, **kwargs):
         params = {
             "name": "New Component With File Params",
             "slug": "new-component-with-file-params",
@@ -73,6 +71,30 @@ class ComponentFileFormatsParamsTest(BaseFileFormatsTest):
             self.assertEqual(response.status_code, 200)
         return response
 
+    def do_create_with_encoding_test(
+        self, encoding_name: str, encoding: str, success: bool = True
+    ) -> None:
+        kwargs = {f"file_format_params_{encoding_name}": encoding}
+        response = self.client_create_component(
+            success,
+            file_format=self.component.file_format,
+            filemask=self.component.filemask,
+            new_base="",
+            new_lang="contact",
+            **kwargs,
+        )
+        if success:
+            self.assertTrue(
+                Component.objects.filter(slug="new-component-with-file-params").exists()
+            )
+        else:
+            self.assertContains(response, "Could not parse")
+            self.assertFalse(
+                Component.objects.filter(slug="new-component-with-file-params").exists()
+            )
+
+
+class ComponentFileFormatsParamsTest(BaseFileFormatsTest):
     def get_new_component(
         self, slug: str = "new-component-with-file-params"
     ) -> Component:
@@ -426,3 +448,55 @@ class GettextParamsTest(BaseFileFormatsTest):
             set(BilingualUpdateMixin.get_msgmerge_args(self.component)),
             {"--no-fuzzy-matching", "--no-location", "--no-wrap"},
         )
+
+
+class StringsParamsTest(BaseFileFormatsTest):
+    def create_component(self):
+        return self.create_iphone()
+
+    def test_encoding_param(self):
+        self.do_create_with_encoding_test("strings_encoding", "utf-8", success=False)
+        self.do_create_with_encoding_test("strings_encoding", "utf-16", success=True)
+
+    def test_new_file_content(self):
+        from weblate.formats.ttkit import StringsFormat
+
+        self.assertNotEqual(
+            StringsFormat.get_new_file_content("utf-8"),
+            StringsFormat.get_new_file_content("utf-16"),
+        )
+
+
+class JavaPropertiesTest(BaseFileFormatsTest):
+    def create_component(self):
+        return self.create_java()
+
+    def test_encoding_param(self):
+        self.do_create_with_encoding_test(
+            "properties_encoding", "utf-16", success=False
+        )
+        self.do_create_with_encoding_test(
+            "properties_encoding", "iso-8859-1", success=True
+        )
+
+    def test_encoding_param_utf8(self):
+        # Java properties need to be ISO 8859-1, but Translate Toolkit converts
+        # them to UTF-8.
+        self.do_create_with_encoding_test("properties_encoding", "utf-8", success=True)
+
+
+class CSVParamsTest(BaseFileFormatsTest):
+    def create_component(self) -> Component:
+        return self.create_csv_mono()
+
+    def test_encoding_param(self):
+        # both "auto" and "utf-8" are valid for the test CSV files
+        self.do_create_with_encoding_test("csv_encoding", "utf-8", success=True)
+
+
+class CSVSimpleParamsTest(BaseFileFormatsTest):
+    def create_component(self) -> Component:
+        return self.create_csv()
+
+    def test_encoding_param(self):
+        self.do_create_with_encoding_test("csv_simple_encoding", "utf-8", success=True)