+++ weblate/trans/tests/utils.py
@@ -322,7 +322,10 @@ class RepoTestMixin:
 
     def create_iphone(self, **kwargs) -> Component:
         return self._create_component(
-            "strings", "iphone/*.lproj/Localizable.strings", **kwargs
+            "strings",
+            "iphone/*.lproj/Localizable.strings",
+            file_format_params={"strings_encoding": "utf-16"},
+            **kwargs,
         )
 
     def create_android(self, suffix="", **kwargs) -> Component:
@@ -381,7 +384,9 @@ class RepoTestMixin:
         return self._create_component("csv", "tsv/*.txt")
 
     def create_csv(self) -> Component:
-        return self._create_component("csv", "csv/*.txt")
+        return self._create_component(
+            "csv", "csv/*.txt", file_format_params={"csv_encoding": "auto"}
+        )
 
     def create_csv_mono(self) -> Component:
         return self._create_component("csv", "csv-mono/*.csv", "csv-mono/en.csv")
@@ -389,11 +394,13 @@ class RepoTestMixin:
     def create_php_mono(self) -> Component:
         return self._create_component("php", "php-mono/*.php", "php-mono/en.php")
 
-    def create_java(self) -> Component:
+    def create_java(self, **kwargs) -> Component:
         return self._create_component(
             "properties",
             "java/swing_messages_*.properties",
             "java/swing_messages.properties",
+            file_format_params={"properties_encoding": "iso-8859-1"},
+            **kwargs,
         )
 
     def create_xliff(self, name="default", **kwargs) -> Component: